+++ CHANGELOG.rst
@@ -106,6 +106,7 @@ Previously-deprecated API have been removed, including:
 - Remove ``__version__``, ``__parsed_version__``, and ``__version_info__`` attributes which were deprecated in 3.21.0.
 
 3.24.0 (unreleased)
+3.24.0 (2025-01-06)
 *******************
 
 Features:
@@ -116,6 +117,16 @@ Features:
 Bug fixes:
 
 - Typing: Fix type hint for ``nested`` parameter of `Nested <marshmallow.fields.Nested>` (:pr:`2721`).
+
+Deprecations:
+
+- Custom validators should raise a `ValidationError <marshmallow.exceptions.ValidationError>` for invalid values. 
+  Returning `False`` is no longer supported .
+- Deprecate ``context`` parameter of `Schema <marshmallow.schema.Schema>` (:issue:`1826`).
+  Use `contextVars.ContextVar` to pass context data instead.
+- `Field <marshmallow.fields.Field>`, `Mapping <marshmallow.fields.Mapping>`,
+  and `Number <marshmallow.fields.Number>` should no longer be used as fields within schemas.
+  Use their subclasses instead.
 
 
 3.23.3 (2025-01-03)